made refactoring code(added services)
